Fixes disqus IDs to be (more) unqiue across a site.
Fixes issue #22 by prefixing the article's slug with year and month. This is still not guaranteed to be unique, but anything other than the URL hardly is, and the point of not using the URL is to protect from future changes of the URL (such as from moving to a different URL layout for the site). The date of a post, at least up to year and month, is hopefully highly unlikely to change once comments start accumulating (if it does prior to the first comment, that doesn't matter, really).