Allows a page's menu label be different from its title.
The title of a page by default is used both for printing the title of the
page's content, and, if pages in menu is enabled, for providing the label
of the menu item. Sometime one may want those to be different, though, such
as having a short (e.g., single word) menu item label, and a longer page title.
This change allows that by adding a Menulabel metadata attribute to the page's
header (Menulabel: in markdown, :Menulabel: in rst).