Don't display SITENAME in banner if HIDE_SITENAME is set
It's probably safe to assume that if people don't want to see SITENAME
in the title, they wouldn't want to see it in the banner, either. Add
an {% if %} test so that SITENAME is only displayed if HIDE_SITENAME
isn't set.

In addition, wrap the whole div in an {% if %} expression such that
unless we want to display SITENAME or BANNER_SUBTITLE, we don't
generate the div at all. The div gets a background from CSS, and would
appear as a square blot on a banner that would be assumed to be
standing on its own.